Move ledger-specific code to indexer_common and consolidate dust processing
