+++ chain-indexer/src/domain.rs
@@ -17,7 +17,6 @@ pub mod storage;
 mod block;
 mod contract_action;
 mod dust;
-mod dust_processing;
 mod ledger_state;
 mod transaction;
 mod transaction_fees;
@@ -25,7 +24,6 @@ mod transaction_fees;
 pub use block::*;
 pub use contract_action::*;
 pub use dust::*;
-pub use dust_processing::*;
 pub use ledger_state::*;
 pub use transaction::*;
 pub use transaction_fees::*;
