+++ chain-indexer/src/domain/dust.rs
@@ -12,11 +12,13 @@
 // limitations under the License.
 
 pub use indexer_common::domain::dust::{
-    DustEvent, DustEventDetails, DustEventType, DustGenerationInfo, DustParameters,
-    QualifiedDustOutput,
+    DustEvent, DustEventDetails, DustEventType, DustGenerationInfo, DustMerklePathEntry,
+    DustParameters, QualifiedDustOutput,
 };
 
-use indexer_common::domain::{CardanoStakeKey, DustAddress, DustUtxoId};
+use indexer_common::domain::{
+    CardanoStakeKey, DustAddress, DustCommitment, DustNullifier, DustUtxoId,
+};
 use thiserror::Error;
 
 /// Domain representation of DUST registration events from the NativeTokenObservation pallet.
@@ -59,4 +61,118 @@ pub enum DustProcessingError {
 
     #[error("DUST generation info not found for index {0}")]
     GenerationInfoNotFound(u64),
+}
+
+/// Processed DUST events ready for persistence.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct ProcessedDustEvents {
+    pub generations: Vec<DustGeneration>,
+    pub utxos: Vec<DustUtxo>,
+    pub merkle_tree_updates: Vec<DustMerkleTreeUpdate>,
+    pub spends: Vec<DustSpend>,
+    pub dtime_update: Option<DustDtimeUpdate>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct DustGeneration {
+    pub generation_info: DustGenerationInfo,
+    pub generation_index: u64,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct DustUtxo {
+    pub output: QualifiedDustOutput,
+    pub generation_info_id: i64,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct DustMerkleTreeUpdate {
+    pub generation_index: u64,
+    pub merkle_path: Vec<DustMerklePathEntry>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct DustSpend {
+    pub commitment: DustCommitment,
+    pub nullifier: DustNullifier,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct DustDtimeUpdate {
+    pub dtime: u64,
+    pub generation_index: u64,
+}
+
+/// Extract operations from DUST events.
+/// This processing happens in the domain layer.
+pub fn extract_dust_operations(dust_events: &[DustEvent]) -> ProcessedDustEvents {
+    let mut result = ProcessedDustEvents {
+        generations: Vec::new(),
+        utxos: Vec::new(),
+        merkle_tree_updates: Vec::new(),
+        spends: Vec::new(),
+        dtime_update: None,
+    };
+
+    let mut generation_dtime_and_index = None;
+
+    for dust_event in dust_events {
+        match &dust_event.event_details {
+            DustEventDetails::DustInitialUtxo {
+                output,
+                generation_info,
+                generation_index,
+            } => {
+                result.generations.push(DustGeneration {
+                    generation_info: *generation_info,
+                    generation_index: *generation_index,
+                });
+                // Note: generation_info_id will be determined during storage.
+                result.utxos.push(DustUtxo {
+                    output: *output,
+                    generation_info_id: 0, // Placeholder, will be set during actual save.
+                });
+            }
+
+            DustEventDetails::DustGenerationDtimeUpdate {
+                generation_info,
+                generation_index,
+                merkle_path,
+            } => {
+                generation_dtime_and_index = Some((generation_info.dtime, *generation_index));
+                result.merkle_tree_updates.push(DustMerkleTreeUpdate {
+                    generation_index: *generation_index,
+                    merkle_path: merkle_path.clone(),
+                });
+            }
+
+            DustEventDetails::DustSpendProcessed {
+                commitment,
+                nullifier,
+                ..
+            } => {
+                result.spends.push(DustSpend {
+                    commitment: *commitment,
+                    nullifier: *nullifier,
+                });
+            }
+
+            // Registration events are handled at block level.
+            DustEventDetails::DustRegistration { .. }
+            | DustEventDetails::DustDeregistration { .. }
+            | DustEventDetails::DustMappingAdded { .. }
+            | DustEventDetails::DustMappingRemoved { .. } => {
+                // Intentionally empty - already handled at block level.
+            }
+        }
+    }
+
+    if let Some((dtime, index)) = generation_dtime_and_index {
+        result.dtime_update = Some(DustDtimeUpdate {
+            dtime,
+            generation_index: index,
+        });
+    }
+
+    result
 }
